cpufreq: Resolve CPUFREQ_NOTIFY issue
When the cpufreq_stat receives the frequency policy notification,
it checks if the frequency table and stats table are present for the cpu.
If both are present, it returns with -EBUSY error to the notification.
Due to this, none of the other registered modules receive this
notification.

If the stat table and sysfs nodes are already created, return 0 instead
of returning EBUSY error.

Signed-off-by: Archana Sathyakumar <asathyak@codeaurora.org>
Signed-off-by: Ad!thya R <gh0strider.2k18.reborn@gmail.com>
Signed-off-by: SakilMondal <mondalsakil@gmail.com>